+++ src/netty_tests/mod.rs
@@ -1,0 +1,2 @@
+mod client_ping;
+mod client_jump;